Move correlation services to infrastructure layer
