+++ source/Energinet.DataHub.MeteringPoints.EntryPoints.Processing/QueueSubscriber.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Application;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Correlation;
 using Energinet.DataHub.MeteringPoints.Infrastructure.Transport;
 using MediatR;
 using Microsoft.Azure.Functions.Worker;
